+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * `19/04/24`: init the project, publish 1.0.
+* `19/05/27`: fix the bug: when the activity is finishing, the anim stop. publish 1.1.